Hitalent/requirements.txt
django
django-allauth
whitenoise
gunicorn
django-cors-headers
djangorestframework
Pillow